pyproject.toml
# CLAP - Chat Local And Persistent
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "clap-llm"
dynamic = ["version"]
description = "CLAP: Chat Local And Persistent - Local LLM conversation tool with Ollama"
readme = "README.md"
license = {text = "GPL-3.0-or-later"}
requires-python = ">=3.10"
authors = [{name = "Frederick", email = "wedonotuse@outlook.com"}]
keywords = ["llm", "ollama", "rag", "local-ai", "chatbot", "pyside6"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "PySide6>=6.5.0",
    "ollama>=0.1.0",
    "mistune>=3.0.0",
    "Pillow>=9.0.0",
]

[project.optional-dependencies]
rag = ["pypdf>=4.0.0", "python-docx>=0.8.11"]
web = ["flask>=2.3.0"]
cli = ["rich>=13.0.0"]
dev = ["pytest>=7.3.0", "ruff>=0.0.270", "build>=1.0.0"]
all = ["clap-llm[rag,web,cli,dev]"]

[project.urls]
Homepage = "https://github.com/cycleuser/CLAP"
Repository = "https://github.com/cycleuser/CLAP.git"

[project.scripts]
clap = "clap:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["clap", "clap.core", "clap.gui", "clap.cli", "clap.web", "clap.utils"]

[tool.setuptools.dynamic]
version = {attr = "clap.__version__"}

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP"]
ignore = ["E501", "B008"]